implement toggle feature for favorites

+++ app/src/main/java/gr/sportsbook/ui/model/Category.kt
@@ -3,13 +3,13 @@ package gr.sportsbook.ui.model
 // Sample data class
 data class Category(
     val title: String,
-    val games: List<GameUiModel> // This should match what you assign in ViewModel
+    val games: List<GameUiModel>
 )
 data class GameUiModel(
     val eventId: String,
     val team1: String,
     val team2: String,
-    val eventStartingTime: Long,
+    val eventStartingTime: Int,
     var isFavorite: Boolean = false
 )
